fix: circular dep b/w default.datatype.ts → utils.ts → value-accessor.ts → location.datatype.ts
