+++ src/app/child-dev-project/notes/notes-related-to-entity/notes-related-to-entity.component.ts
@@ -7,7 +7,7 @@ import { Entity } from "../../../core/entity/model/entity";
 import { FilterService } from "../../../core/filter/filter.service";
 import { ChildSchoolRelation } from "../../children/model/childSchoolRelation";
 import { EntityDatatype } from "../../../core/basic-datatypes/entity/entity.datatype";
-import { asArray } from "../../../utils/utils";
+import { asArray } from "app/utils/asArray";
 import { EntitiesTableComponent } from "../../../core/common-components/entities-table/entities-table.component";
 import { EntityMapperService } from "../../../core/entity/entity-mapper/entity-mapper.service";
 import { FormFieldConfig } from "../../../core/common-components/entity-form/FormConfig";
