+++ src/app/core/basic-datatypes/configurable-enum/configurable-enum-ordering.ts
@@ -1,4 +1,7 @@
-import { ConfigurableEnumConfig, ConfigurableEnumValue, HasOrdinal } from "./configurable-enum-types";
+import {
+  ConfigurableEnumConfig,
+  ConfigurableEnumValue,
+} from "./configurable-enum.interface";
 import { isObject } from "lodash-es";
 
 export namespace Ordering {
@@ -15,24 +18,32 @@ export namespace Ordering {
    * and thus have a notion of one element being 'greater than' or 'less than' to another element. The interpretation
    * of 'greater' or 'less' than is dependent on the concrete enum.
    */
+  export interface HasOrdinal {
+    _ordinal?: number;
+  }
+
   export function hasOrdinalValue(value: any): value is HasOrdinal {
     return isObject(value) && "_ordinal" in value;
   }
 
-  /**
-   * Extends ConfigurableEnumValue with HasOrdinal for ordering.
-   */
-  export type EnumValue<T extends ConfigurableEnumValue = ConfigurableEnumValue> = T & HasOrdinal;
+  export type EnumValue<
+    T extends ConfigurableEnumValue = ConfigurableEnumValue,
+  > = T & HasOrdinal;
+
+  export type Config<T extends ConfigurableEnumValue> = ConfigurableEnumConfig<
+    T & HasOrdinal
+  >;
 
   /**
-   * Configuration type for ordered enums.
-   */
-  export type Config<T extends ConfigurableEnumValue> = ConfigurableEnumConfig<T & HasOrdinal>;
-
-  /**
-   * Assigns ordinal values to an array of configurable enum values.
-   * 
-   * @param values The values to impose a total ordering on.
+   * Extends an array of configurable enum values with an ordinal value so that the
+   * ordinal value of each enum value matches the position of the element in the array.
+   *
+   * Note that this should be used 'early' in the pipeline, i.e. when the `ConfigurableEnumConfig` is just
+   * created. It is technically not incorrect to use this function on any array that contains configurable
+   * enum values, but it doesn't make sense to apply this function to 'scrambled' arrays, i.e. arrays where the
+   * ordering of values is _not_ the natural order of these values.
+   *
+   * @param values The values to impose a total ordering on
    */
   export function imposeTotalOrdering<T extends ConfigurableEnumValue>(
     values: ConfigurableEnumConfig<T>,
