+++ src/app/core/basic-datatypes/entity/display-entity/display-entity.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from "@angular/core";
 import { ViewDirective } from "../../../entity/default-datatype/view.directive";
 import { DynamicComponent } from "../../../config/dynamic-components/dynamic-component.decorator";
 import { EntityBlockComponent } from "../entity-block/entity-block.component";
-import { asArray } from "../../../../utils/utils";
+import { asArray } from "app/utils/asArray";
 
 @DynamicComponent("DisplayEntity")
 @Component({
