+++ src/app/core/default-values/inherited-value.service.ts
@@ -10,7 +10,7 @@ import {
 import { EntityMapperService } from "../entity/entity-mapper/entity-mapper.service";
 import { DefaultValueConfig } from "../entity/schema/default-value-config";
 import { DefaultValueHint } from "./default-value.service";
-import { asArray } from "../../utils/utils";
+import { asArray } from "app/utils/asArray";
 import { FormFieldConfig } from "../common-components/entity-form/FormConfig";
 
 /**
