+++ src/app/features/location/GeoLocation.ts
@@ -1,0 +1,10 @@
+import { GeoResult } from "./geo.service";
+
+/**
+ * A location both as custom string and an optional geo location lookup.
+ */
+
+export interface GeoLocation {
+  locationString?: string;
+  geoLookup?: GeoResult;
+}
