+++ src/app/features/location/address-edit/address-edit.component.spec.ts
@@ -11,7 +11,7 @@ import { FontAwesomeTestingModule } from "@fortawesome/angular-fontawesome/testi
 import { ConfirmationDialogService } from "../../../core/common-components/confirmation-dialog/confirmation-dialog.service";
 import { of } from "rxjs";
 import { NoopAnimationsModule } from "@angular/platform-browser/animations";
-import { GeoLocation } from "../location.datatype";
+import { GeoLocation } from "../GeoLocation";
 
 describe("AddressEditComponent", () => {
   let component: AddressEditComponent;
