+++ src/app/features/location/location-input/location-input.component.spec.ts
@@ -13,7 +13,7 @@ import {
   MapPopupComponent,
   MapPopupConfig,
 } from "../map-popup/map-popup.component";
-import { GeoLocation } from "../location.datatype";
+import { GeoLocation } from "../GeoLocation";
 
 describe("LocationInputComponent", () => {
   let component: LocationInputComponent;
