+++ src/app/features/location/location-input/location-input.component.ts
@@ -6,7 +6,7 @@ import {
   NgForm,
 } from "@angular/forms";
 import { MatFormFieldControl, MatSuffix } from "@angular/material/form-field";
-import { GeoLocation } from "../location.datatype";
+import { GeoLocation } from "../GeoLocation";
 import { MatInput } from "@angular/material/input";
 import { MatIconButton } from "@angular/material/button";
 import { FaIconComponent } from "@fortawesome/angular-fontawesome";
