+++ src/app/features/location/location.datatype.spec.ts
@@ -1,6 +1,7 @@
 import { TestBed } from "@angular/core/testing";
 import { of } from "rxjs";
-import { GeoLocation, LocationDatatype } from "./location.datatype";
+import { LocationDatatype } from "./location.datatype";
+import { GeoLocation } from "./GeoLocation";
 import { GeoResult, GeoService } from "./geo.service";
 
 describe("Schema data type: location", () => {
