+++ src/app/features/location/map/map.component.spec.ts
@@ -14,7 +14,7 @@ import { MatDialog } from "@angular/material/dialog";
 import { MapPopupConfig } from "../map-popup/map-popup.component";
 import { FontAwesomeTestingModule } from "@fortawesome/angular-fontawesome/testing";
 import { EMPTY, of, Subject } from "rxjs";
-import { GeoLocation } from "../location.datatype";
+import { GeoLocation } from "../GeoLocation";
 import { TestEntity } from "../../../utils/test-utils/TestEntity";
 
 describe("MapComponent", () => {
