+++ src/app/utils/utils.ts
@@ -157,10 +157,4 @@ export function serviceProvider<T>(
   };
 }
 
-/**
- * Convert wrap a value in an array if it is not already an array.
- * @param x
- */
-export function asArray<T>(x: T | T[]): T[] {
-  return Array.isArray(x) ? x : [x];
-}
+
